+++ lib/txdx/txdx_syntax.dart
@@ -1,4 +1,5 @@
 import 'package:jiffy/jiffy.dart';
+import 'package:clock/clock.dart';
 
 class TxDxSyntax {
   static RegExp contextsRegExp = RegExp(r'(?:\s+|^)@[^\s]+');
@@ -35,26 +36,57 @@ class TxDxSyntax {
   static Map<String, String> getTags(String text) {
     final tags = _getMatchedPairs(tagsRegExp, text);
 
+    String _formattedDate(DateTime dateTime) {
+      return Jiffy(dateTime).format('yyyy-MM-dd');
+    }
+
+    String _futureFormattedDate(int dayDelta) {
+      final now = clock.now();
+      final targetDate = DateTime(now.year, now.month, now.day + dayDelta);
+      return _formattedDate(targetDate);
+    }
+
+    String _futureFormattedWeekDate(int targetWd) {
+      final now = clock.now();
+      final nowWd = now.weekday;
+      var dayDelta = (nowWd < targetWd)
+          ? targetWd - nowWd
+          : (targetWd + 7) - nowWd;
+
+      return _futureFormattedDate(dayDelta);
+    }
+
     // Find/replace the 'due' tag if it exists.
     if (tags.containsKey('due')) {
       final value = tags['due']!.toLowerCase();
-      final now = DateTime.now();
+
 
       if (todoTxtDate.hasMatch(value)) {
         tags['due'] = Jiffy(value).format('yyyy-MM-dd');
       } else if (value == 'today') {
-        final today = DateTime(now.year, now.month, now.day);
-        tags['due'] = Jiffy(today).format('yyyy-MM-dd');
+        tags['due'] = _futureFormattedDate(0);
       } else if (value == 'tomorrow') {
-        final tomorrow = DateTime(now.year, now.month, now.day + 1);
-        tags['due'] = Jiffy(tomorrow).format('yyyy-MM-dd');
+        tags['due'] =_futureFormattedDate(1);
       } else if (value == 'yesterday') {
-        final yesterday = DateTime(now.year, now.month, now.day - 1);
-        tags['due'] = Jiffy(yesterday).format('yyyy-MM-dd');
+        tags['due'] =_futureFormattedDate(-1);
+      } else if (value == 'mon' || value == 'monday') {
+        tags['due'] =_futureFormattedWeekDate(DateTime.monday);
+      } else if (value == 'tue' || value == 'tuesday') {
+        tags['due'] =_futureFormattedWeekDate(DateTime.tuesday);
+      } else if (value == 'wed' || value == 'wednesday') {
+        tags['due'] =_futureFormattedWeekDate(DateTime.wednesday);
+      } else if (value == 'thu' || value == 'thursday') {
+        tags['due'] =_futureFormattedWeekDate(DateTime.thursday);
+      } else if (value == 'fri' || value == 'friday') {
+        tags['due'] =_futureFormattedWeekDate(DateTime.friday);
+      } else if (value == 'sat' || value == 'saturday') {
+        tags['due'] =_futureFormattedWeekDate(DateTime.saturday);
+      } else if (value == 'sun' || value == 'sunday') {
+        tags['due'] =_futureFormattedWeekDate(DateTime.sunday);
       } else {
         var match = incDaysRegExp.firstMatch(value);
         if (match != null) {
-          var dueOn = now;
+          var dueOn = clock.now();
           final yearVal = match.group(1);
           final monthVal = match.group(2);
           final weekVal = match.group(3);
@@ -77,7 +109,7 @@ class TxDxSyntax {
             dueOn = Jiffy(dueOn).add(days: value).dateTime;
           }
 
-          tags['due'] = Jiffy(dueOn).format('yyyy-MM-dd');
+          tags['due'] = _formattedDate(dueOn);
         }
       }
     }